rework / simplify interface.
Was too broad in scope: I removed notions of ook knowing things
about particular file formats.  This removes most notions of types
and number of components and bit widths and so on; now ook just
thinks of data as indices it needs to figure out.

Ook's scope is thus reduced to bricking the data.  All other
concerns are left to the user.

Specifically:

. read (which is really an open) takes a volume size and a
bricking size.  It does not / cannot figure this out on its own.
It assumes the user knows what they're doing.

. type / n-components / width functions have been removed.

. ookbrickout was renamed to ookwrite.  we don't write anything
except bricks.

. a notion of a 'reader' was introduced.  this allows users to
account for pecularities of their file format.  for example, a
reader for a nrrd file might add in the offset of the header
before reading the data.