implement writing as well, new test program.
abstract out 'ookbrick' into 'srcop', which takes a function
pointer so it can control in which direction the data are going.
reimplement ookbrick in terms of srcop, and implement ookwrite
by just passing in a 'write' pointer.

ookwrite doesn't need a length parameter: the length is implicit
based on the brick passed in.

the new test program is just a simple thresholding app: it writes
a new binary volume with a voxel 'on' if it passes the threshold
operation.